Recreate entire reporter when the sending thread dies
