+++ lib/signalfx/tracing/reporter/auto_reviving_async_reporter.rb
@@ -5,7 +5,7 @@ require 'jaeger/client/async_reporter'
 # process that includes the tracer, the sender thread is lost.
 #
 # This checks for the thread's before pushing in a span to the buffer.
-# If it doesn't exist, it creates a new thread.
+# If it doesn't exist, it tells the instrumenter module to crate a new reporter.
 #
 # If you have control over hooking into fork events, signalfx/tracing/async_reporter
 # and reviving it should be preferred to avoid an unnecessary check with every
@@ -13,9 +13,21 @@ require 'jaeger/client/async_reporter'
 
 module SignalFx
   module Tracing
-    class AutoRevivingAsyncReporter < SignalFx::Tracing::AsyncReporter
+    class AutoRevivingAsyncReporter < ::Jaeger::Client::AsyncReporter
+      def initialize(sender, flush_interval)
+        @flush_interval = flush_interval
+        @poll_thread = Thread.new do
+          loop do
+            flush
+            sleep(@flush_interval)
+          end
+        end
+
+        super(sender)
+      end
+
       def report(span)
-        revive_poll_thread if !@poll_thread
+        ::SignalFx::Tracing::Instrumenter.revive if !@poll_thread
         super
       end
     end
